// src/types/Student.ts
export interface Student {
  id?: number;

  fname: string;
  lname: string;
  gender: string;
  dateofbirth: string;
  branch: string;
  dateofjoining: string;
  graduation: string;
  fathername: string;
  mothername: string;
  mobile: string;
  email: string;
  aadhar: string;
  photo: string;
  ssc: string;
  inter: string;
}